FPGA: Add echo_ram, which reads bytes then responds.
This test reads a large sequence of bytes then returns the same message.
This was intended as a test of the block ram module within the ice40.
Using registers directly was simpler to implement, and still fits
on the ice40.

This HDL is similar to what will be needed in the chiplab. At
the beginning of a run, memory will be copied form the host to the
FPGA. During execution, the fpga will present the target chip
with whatever bits are requested via the address bus.